+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added support for `wasm32-unknown-unknown` target using Web APIs
 - Return inner error as source in `RequestError::Endpoint` variant
 - Introduced new `api::ResponseError` type
 - Exported `api::Response` type